feat: add test detail overview and ability to start and end tests
