+++ proctor/src/router/index.ts
@@ -1,8 +1,21 @@
 import { createRouter, createWebHistory } from 'vue-router'
+import HomeView from '@/views/HomeView.vue'
+import TestView from '@/views/TestView.vue'
 
 const router = createRouter({
   history: createWebHistory(import.meta.env.BASE_URL),
-  routes: [],
+  routes: [
+    {
+      path: '/',
+      name: 'home',
+      component: HomeView,
+    },
+    {
+      path: '/test/:id',
+      name: 'test',
+      component: TestView,
+    },
+  ],
 })
 
 export default router
