+++ server/src/main/kotlin/at/ac/htlleonding/franklyn/test/entity/dto/TestListDTO.kt
@@ -8,6 +8,6 @@ data class TestListDTO(
     val testAccountPrefix: String,
     val start: LocalDateTime?,
     val end: LocalDateTime?,
-    val teacherId: Long?,
+    val teacherId: Long?
 )
